Update DFS.java
Requirements:
visited array
stack
Time Complexity :O(V+E)


We start pushing the data into the stack, marking it visited in the visited array and also printing it
Now look for nearnest nodes using the edges and select another node and push that node also in the stack. Keep on doing it till adjacents nodes are not visited. If alll adjacent nodes are visited pop out the element from stack and continue the above process.
When the stack is empty the algorithm stops.